TP8 - Juan se hartó del admin

+++ 8-tp/operations.c
@@ -2,33 +2,33 @@
 
 
 
-static float calc_sen(float a);
-static float calc_factorial(int a);
-static float calc_potencia(float a, int b);
+static double calc_sen(double a);
+static double calc_factorial(int a);
+static double calc_potencia(double a, int b);
 
 
 
-float suma(float a, float b){
+double suma(double a, double b){
     return a+b;	
 }
 
-float resta(float a, float b){
+double resta(double a, double b){
     return a-b;
 }
 
-float division(float a, float b){
+double division(double a, double b){
     return a/b;
 }
 
-float multiplicacion(float a, float b){
+double multiplicacion(double a, double b){
     return a*b;
 }
 
-float potencia(float a, float b) {
+double potencia(double a, double b) {
     return calc_potencia(a, (int) b);
 }
 
-static float calc_potencia(float a, int b){
+static double calc_potencia(double a, int b){
     int i, ans = 1;
     for (i=0; i<b; i++){
         ans *= a;
@@ -36,13 +36,13 @@ static float calc_potencia(float a, int b){
     return ans;
 }
 
-float sen(float a, float b) {
+double sen(double a, double b) {
     return calc_sen(a);
 }
 
 #define N_TERM 8
-static float calc_sen(float a){
-    float ans, termino;
+static double calc_sen(double a){
+    double ans, termino;
     for (int i=0; i<N_TERM; i++) {
         int exp = 2*i + 1;                            //esto calcula el exponente y el argumento del factorial
         termino = calc_potencia(a, exp) / calc_factorial(exp);     //calcula el valor absoluto de cada termino
@@ -56,17 +56,16 @@ static float calc_sen(float a){
     return ans;
 }
 
-float factorial(float a, float b) {
+double factorial(double a, double b) {
     return calc_factorial((int) a);
 }
 
-static float calc_factorial(int a){
+static double calc_factorial(int a){
     int ans = 1;                                      //comienza ans en 1 por si quiere calcular 0!
     int term = a;
     
     if(a < 0) {
         return 0.0/0.0;
-        
     }
 
     while (term > 1){                                //evita mutiplicar por cero y se detiene en term == 1
